[chore] 비동기 로직도 기존 CouponController, CouponFacade.kt 에 포함
